requirements.txt
fastapi
uvicorn
pytest
ruff
python-json5
gymnasium
stable-baselines3